# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "x402 API Portfolio",
    "version": "1.0.0",
    "description": "Five machine-native REST services protected by x402 payments on Base USDC. Every operation below is paid and returns HTTP 402 with x402 payment requirements before a successful paid retry."
  },
  "servers": [{"url": "https://x402-api-91r3.vercel.app", "description": "Production"}],
  "security": [{"x402Payment": []}],
  "paths": {
...
    "/api/scraped-data": {
      "get": {
        "operationId": "scrapeData",
        "summary": "Scrape a target URL",
        "parameters": [{"name": "target", "in": "query", "required": false, "schema": {"type": "string", "format": "uri", "default": "https://example.com"}}],
        "responses": {"200": {"$ref": "#/components/responses/ScrapeResponse"}, "402": {"$ref": "#/components/responses/PaymentRequired"}}
      }
    }
  },
  "components": {
    "securitySchemes": {"x402Payment": {"type": "apiKey", "in": "header", "name": "X-PAYMENT", "description": "Base64-encoded x402 payment payload sent on the retry request."}},
    "headers": {
      "PaymentRequired": {"description": "Base64-encoded x402 payment requirements. Includes network, USDC asset, payTo, resource, and amount.", "schema": {"type": "string"}},
...
    },
...
    "responses": {
      "PaymentRequired": {"description": "HTTP 402 payment challenge. Sign an EIP-3009 USDC transferWithAuthorization based on the requirements, send it as X-PAYMENT, and retry. Successful responses include PAYMENT-RESPONSE.", "headers": {"PAYMENT-REQUIRED": {"$ref": "#/components/headers/PaymentRequired"}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaymentChallenge"}}}},
...
      "ScrapeResponse": {"description": "Paid scraped data", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScrapeResult"}}}}
    },
    "schemas": {
      "EvmAddress": {"type": "string", "pattern": "^0x[a-fA-F0-9]{40}$", "example": "0x0000000000000000000000000000000000000001"},
      "PaymentChallenge": {"type": "object", "required": ["x402Version", "accepts"], "properties": {"x402Version": {"type": "integer", "example": 1}, "accepts": {"type": "array", "items": {"$ref": "#/components/schemas/PaymentRequirement"}}}},
      "PaymentRequirement": {"type": "object", "required": ["scheme", "network", "asset", "payTo", "maxAmountRequired", "resource"], "properties": {"scheme": {"type": "string", "example": "exact"}, "network": {"type": "string", "example": "base"}, "asset": {"type": "string", "example": "USDC"}, "payTo": {"$ref": "#/components/schemas/EvmAddress"}, "maxAmountRequired": {"type": "string", "description": "USDC atomic units with 6 decimals"}, "resource": {"type": "string"}, "description": {"type": "string"}}},
...
      "ScrapeResult": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"type": "object", "additionalProperties": true}}}
    }
  }
}